Move to Chromium 44.0.2403.107.
Not much to report other than that it's the second Linux stable release
and there are a few changes to make the code build with changes in
Chromium.

+++ sysapps/raw_socket/udp_socket_object.cc
@@ -17,7 +17,7 @@ using namespace xwalk::jsapi::raw_socket; // NOLINT
 
 namespace {
 
-const unsigned kBufferSize = 4096;
+const size_t kBufferSize = 4096;
 
 }  // namespace
